+++ examples/ex2/main.cpp
@@ -111,7 +111,7 @@ int main() {
     }
 
     // integrate
-    const long Nmc = 1000000;
+    const int Nmc = 1000000;
     double average[mci.getNObsDim()];
     double error[mci.getNObsDim()];
